Require only the types that are used
When using it with a preload function you don't have access to the electron in the `window` variable, however you can make an API which have the `send` and `on` functions, and you can pass that into the lib. I also like to import from libraries what is needed like `ipcMain` instead of the whole electron.